d9a58bb1 COMP: Fix particle reader warnings
105287ee COMP: Fix AMR resampler filter warnings
1fc7f61a COMP: Fix AMR slice filter warnings
ecd3ae58 COMP: Fix AMRBaseReader compiler warnings
b23f4c28 COMP: Fix minor compiler warnings
f0816b60 COMP: Conditional compilation of AMR VolumeMapper
7df7ebcf ENH: Updated AMR volume mapper for VTK 6 updates
00c6e363 ENH: Rename SetInput to SetInputData
91ee8646 ENH: Fix PrintSelf-Rendering ivar defects
934d007a ENH: Fix PrintSelf-Filtering ivar defects
41e08200 ENH: Fix PrintSelf-Common ivar defects

+++ AMR/vtkAMRSliceFilter.cxx
@@ -611,7 +611,7 @@ int vtkAMRSliceFilter::RequestUpdateExtent(
     // Send upstream request for higher resolution
     inInfo->Set(
      vtkCompositeDataPipeline::UPDATE_COMPOSITE_INDICES(),
-     &this->blocksToLoad[0], this->blocksToLoad.size() );
+     &this->blocksToLoad[0], static_cast<int>(this->blocksToLoad.size()));
     }
 
   return 1;
